Remove Python 3.9 support and refactor models
- Update pyproject.toml to require Python >=3.10
- Update GitHub Actions workflow to test on Python 3.10+
- Refactor models into separate modules for better organization
- Update imports and type annotations throughout codebase
